backend/requirements.txt
Flask == 1.1.2
flask-cors==3.0.8
Pillow == 7.2.0
colormath == 3.0.0
colour-science==0.3.15
imutils == 0.5.3
numpy == 1.19.1
opencv-python == 4.1.2.30
pandas == 1.1.0
scikit-learn == 0.22.2
scipy == 1.5.2
tensorflow == 1.15.0
torch == 1.3.1
torchvision == 0.4.2
tqdm == 4.48.0
pymongo==3.5.1
